Remove DataType in from_arrow

+++ lonboard/_geoarrow/extension_types.py
@@ -157,14 +157,9 @@ def offsets_to_arrow(
 ) -> Sequence[Array]:
     # Shapely produces int64 offset arrays. We downcast those to int32 if possible
     if any(offset_arr[-1] >= np.iinfo(np.int32).max for offset_arr in offsets):
-        return [
-            Array.from_numpy(offset_arr, DataType.int64()) for offset_arr in offsets
-        ]
+        return [Array.from_numpy(offset_arr) for offset_arr in offsets]
 
-    return [
-        Array.from_numpy(offset_arr.astype(np.int32), DataType.int32())
-        for offset_arr in offsets
-    ]
+    return [Array.from_numpy(offset_arr.astype(np.int32)) for offset_arr in offsets]
 
 
 def construct_geometry_array(
@@ -195,7 +190,7 @@ def construct_geometry_array(
 
     if geom_type == GeometryType.POINT:
         arrow_coords = fixed_size_list_array(
-            Array.from_numpy(coords.ravel("C"), DataType.float64()), len(dims)
+            Array.from_numpy(coords.ravel("C")), len(dims)
         )
         extension_metadata["ARROW:extension:name"] = "geoarrow.point"
         field = Field(
@@ -210,7 +205,7 @@ def construct_geometry_array(
         assert len(offsets) == 1, "Expected one offsets array"
         (geom_offsets,) = offsets
         arrow_coords = fixed_size_list_array(
-            Array.from_numpy(coords.ravel("C"), DataType.float64()), len(dims)
+            Array.from_numpy(coords.ravel("C")), len(dims)
         )
         arrow_geoms = list_array(geom_offsets, arrow_coords)
         extension_metadata["ARROW:extension:name"] = "geoarrow.linestring"
@@ -226,7 +221,7 @@ def construct_geometry_array(
         assert len(offsets) == 2, "Expected two offsets arrays"
         ring_offsets, geom_offsets = offsets
         arrow_coords = fixed_size_list_array(
-            Array.from_numpy(coords.ravel("C"), DataType.float64()), len(dims)
+            Array.from_numpy(coords.ravel("C")), len(dims)
         )
         arrow_rings = list_array(ring_offsets, arrow_coords)
         arrow_geoms = list_array(geom_offsets, arrow_rings)
@@ -243,7 +238,7 @@ def construct_geometry_array(
         assert len(offsets) == 1, "Expected one offsets array"
         (geom_offsets,) = offsets
         arrow_coords = fixed_size_list_array(
-            Array.from_numpy(coords.ravel("C"), DataType.float64()), len(dims)
+            Array.from_numpy(coords.ravel("C")), len(dims)
         )
         arrow_geoms = list_array(geom_offsets, arrow_coords)
         extension_metadata["ARROW:extension:name"] = "geoarrow.multipoint"
@@ -259,7 +254,7 @@ def construct_geometry_array(
         assert len(offsets) == 2, "Expected two offsets arrays"
         ring_offsets, geom_offsets = offsets
         arrow_coords = fixed_size_list_array(
-            Array.from_numpy(coords.ravel("C"), DataType.float64()), len(dims)
+            Array.from_numpy(coords.ravel("C")), len(dims)
         )
         arrow_rings = list_array(ring_offsets, arrow_coords)
         arrow_geoms = list_array(geom_offsets, arrow_rings)
@@ -276,7 +271,7 @@ def construct_geometry_array(
         assert len(offsets) == 3, "Expected three offsets arrays"
         ring_offsets, polygon_offsets, geom_offsets = offsets
         arrow_coords = fixed_size_list_array(
-            Array.from_numpy(coords.ravel("C"), DataType.float64()), len(dims)
+            Array.from_numpy(coords.ravel("C")), len(dims)
         )
         arrow_rings = list_array(ring_offsets, arrow_coords)
         arrow_polygons = list_array(polygon_offsets, arrow_rings)
