+++ lonboard/_viz.py
@@ -19,7 +19,7 @@ from typing import (
 
 import numpy as np
 from arro3.compute import struct_field
-from arro3.core import Array, ChunkedArray, DataType, Field, Schema, Table
+from arro3.core import Array, ChunkedArray, Field, Schema, Table
 
 from lonboard._compat import check_pandas_version
 from lonboard._constants import EXTENSION_NAME
@@ -434,21 +434,13 @@ def _viz_geoarrow_array(
 
     num_rows = len(array)
     if num_rows <= np.iinfo(np.uint8).max:
-        arange_col = Array.from_numpy(
-            np.arange(num_rows, dtype=np.uint8), DataType.uint8()
-        )
+        arange_col = Array.from_numpy(np.arange(num_rows, dtype=np.uint8))
     elif num_rows <= np.iinfo(np.uint16).max:
-        arange_col = Array.from_numpy(
-            np.arange(num_rows, dtype=np.uint16), DataType.uint16()
-        )
+        arange_col = Array.from_numpy(np.arange(num_rows, dtype=np.uint16))
     elif num_rows <= np.iinfo(np.uint32).max:
-        arange_col = Array.from_numpy(
-            np.arange(num_rows, dtype=np.uint32), DataType.uint32()
-        )
+        arange_col = Array.from_numpy(np.arange(num_rows, dtype=np.uint32))
     else:
-        arange_col = Array.from_numpy(
-            np.arange(num_rows, dtype=np.uint64), DataType.uint64()
-        )
+        arange_col = Array.from_numpy(np.arange(num_rows, dtype=np.uint64))
 
     table = table.append_column("row_index", ChunkedArray([arange_col]))
     return _viz_geoarrow_table(table, **kwargs)
